20260516-2 feat: basic schedule and runner

+++ femtotron/model/base.py
@@ -25,7 +25,7 @@ class BaseCausalLMPipeline(nn.Module):
         shift_logits = logits[..., :-1, :].contiguous()
         shift_labels = labels[..., 1:].contiguous()
         return F.cross_entropy(
-            shift_logits.view(-1, shift_logits.size(-1)),
+            shift_logits.float().view(-1, shift_logits.size(-1)),
             shift_labels.view(-1),
             ignore_index=-100,
         )
